break out dashboard route
