+++ client/src/App.js
@@ -2,6 +2,7 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Home from "./components/Home";
 import SignUp from "./pages/SignUp";
+import Dashboard from "./pages/Dashboard";
 import Footer from "./components/Footer";
 import Header from "./components/Header";
 // import Saved from "./Saved/Saved";
@@ -15,6 +16,7 @@ const App = () => (
       <Switch>
         <Route exact path="/" component={Home} />
         <Route exact path="/signup" component={SignUp} />
+        <Route exact path="/dashboard" component={Dashboard} />
         {/* <Route exact path="/saved" component={Saved} />
         <Route component={NoMatch} /> */}
       </Switch>
